refactor: change file name for forms/form.py & add soft delete
