+++ apps/companies/models.py
@@ -1,8 +1,10 @@
 from django.db import models
 
+from lib.models.soft_delete import SoftDeleteManager, SoftDeletetable
+
 
 # Create your models here.
-class Company(models.Model):
+class Company(SoftDeletetable, models.Model):
     title = models.CharField(max_length=200)
     tel = models.CharField(max_length=15)
     url = models.URLField()
@@ -15,6 +17,8 @@ class Company(models.Model):
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
     deleted_at = models.DateTimeField(default=None, null=True)
+
+    objects = SoftDeleteManager()
 
     class Meta:
         indexes = [
